refactor(home): move components to home feature views
