+++ apps/frontend/src/features/home/views/hero/SocialProof.vue
