+++ apps/frontend/src/features/home/views/newsletter/NewsletterCTA.vue
