+++ apps/frontend/src/features/home/views/todays-pick/TodaysPickHeader.vue
